Convert dashboard frontend to TypeScript (#22)
