+++ dashboard/frontend/index.html
@@ -23,6 +23,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app. :(</noscript>
     <div id="root"></div>
-    <script type="module" src="./src/index.jsx"></script>
+    <script type="module" src="./src/index.tsx"></script>
   </body>
 </html>
