+++ dashboard/frontend/src/components/AISettings.test.tsx
