+++ dashboard/frontend/src/components/AISettings.tsx
