+++ dashboard/frontend/src/components/AdminDashboard.tsx
