+++ dashboard/frontend/src/components/AdminGuildDetailsPage.tsx
