+++ dashboard/frontend/src/components/AdminGuildsPage.tsx
