+++ dashboard/frontend/src/components/AdminLayout.tsx
