+++ dashboard/frontend/src/components/AdminRawDBPage.tsx
