+++ dashboard/frontend/src/components/AdminSidebar.tsx
