+++ dashboard/frontend/src/components/AnalyticsDashboard.tsx
