+++ dashboard/frontend/src/components/BlogManagement.tsx
