+++ dashboard/frontend/src/components/ChannelManagement.tsx
