+++ dashboard/frontend/src/components/DashboardPage.tsx
