+++ dashboard/frontend/src/components/DiscordSelector.tsx
