+++ dashboard/frontend/src/components/GeneralSettings.tsx
