+++ dashboard/frontend/src/components/GuildConfigPage.tsx
