+++ dashboard/frontend/src/components/LandingPage.tsx
