+++ dashboard/frontend/src/components/LoggingSettings.test.tsx
