+++ dashboard/frontend/src/components/LoggingSettings.tsx
