+++ dashboard/frontend/src/components/ModerationSettings.test.tsx
