+++ dashboard/frontend/src/components/ModerationSettings.tsx
