+++ dashboard/frontend/src/components/Navbar.tsx
