+++ dashboard/frontend/src/components/ProtectedRoute.tsx
