+++ dashboard/frontend/src/components/RaidDefenseSettings.tsx
