+++ dashboard/frontend/src/components/RateLimitingSettings.tsx
