+++ dashboard/frontend/src/components/Sidebar.tsx
