--- dashboard/frontend/src/components/ui/switch.jsx
@@ -1,30 +1,0 @@
-import React, { forwardRef } from 'react';
-import { cn } from '../../lib/utils';
-
-export const Switch = forwardRef(({ className, checked, onCheckedChange, disabled, ...props }, ref) => {
-  return (
-    <button
-      type="button"
-      role="switch"
-      aria-checked={checked}
-      disabled={disabled}
-      className={cn(
-        'peer inline-flex h-6 w-11 shrink-0 cursor-pointer items-center rounded-full border-2 border-transparent transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-background disabled:cursor-not-allowed disabled:opacity-50',
-        checked ? 'bg-primary' : 'bg-input',
-        className
-      )}
-      onClick={() => onCheckedChange && onCheckedChange(!checked)}
-      ref={ref}
-      {...props}
-    >
-      <span
-        className={cn(
-          'pointer-events-none block h-5 w-5 rounded-full bg-background shadow-lg ring-0 transition-transform',
-          checked ? 'translate-x-5' : 'translate-x-0'
-        )}
-      />
-    </button>
-  );
-});
-
-Switch.displayName = 'Switch';
