+++ dashboard/frontend/src/components/ui/tabs.tsx
@@ -1,66 +1,83 @@
-"use client"
+import React, { createContext, useContext, useState } from 'react';
+import { cn } from '../../lib/utils';
 
-import * as React from "react"
-import * as TabsPrimitive from "@radix-ui/react-tabs"
+const TabsContext = createContext();
 
-import { cn } from "@/lib/utils"
+export const Tabs = ({ value, onValueChange, children, className, ...props }) => {
+  const [internalValue, setInternalValue] = useState(value);
+  
+  const currentValue = value !== undefined ? value : internalValue;
+  const handleValueChange = onValueChange || setInternalValue;
 
-function Tabs({
-  className,
-  ...props
-}: React.ComponentProps<typeof TabsPrimitive.Root>) {
   return (
-    <TabsPrimitive.Root
-      data-slot="tabs"
-      className={cn("flex flex-col gap-2", className)}
-      {...props}
-    />
-  )
-}
+    <TabsContext.Provider value={{ value: currentValue, onValueChange: handleValueChange }}>
+      <div className={cn('w-full', className)} {...props}>
+        {children}
+      </div>
+    </TabsContext.Provider>
+  );
+};
 
-function TabsList({
-  className,
-  ...props
-}: React.ComponentProps<typeof TabsPrimitive.List>) {
+export const TabsList = ({ children, className, ...props }) => {
   return (
-    <TabsPrimitive.List
-      data-slot="tabs-list"
+    <div
       className={cn(
-        "bg-muted text-muted-foreground inline-flex h-9 w-fit items-center justify-center rounded-lg p-[3px]",
+        'inline-flex h-10 items-center justify-center rounded-md bg-muted p-1 text-muted-foreground',
         className
       )}
       {...props}
-    />
-  )
-}
+    >
+      {children}
+    </div>
+  );
+};
 
-function TabsTrigger({
-  className,
-  ...props
-}: React.ComponentProps<typeof TabsPrimitive.Trigger>) {
+export const TabsTrigger = ({ value, children, className, ...props }) => {
+  const context = useContext(TabsContext);
+  
+  if (!context) {
+    throw new Error('TabsTrigger must be used within Tabs');
+  }
+
+  const isActive = context.value === value;
+
   return (
-    <TabsPrimitive.Trigger
-      data-slot="tabs-trigger"
+    <button
       className={cn(
-        "data-[state=active]:bg-background dark:data-[state=active]:text-foreground focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:outline-ring dark:data-[state=active]:border-input dark:data-[state=active]:bg-input/30 text-foreground dark:text-muted-foreground inline-flex h-[calc(100%-1px)] flex-1 items-center justify-center gap-1.5 rounded-md border border-transparent px-2 py-1 text-sm font-medium whitespace-nowrap transition-[color,box-shadow] focus-visible:ring-[3px] focus-visible:outline-1 disabled:pointer-events-none disabled:opacity-50 data-[state=active]:shadow-sm [&_svg]:pointer-events-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4",
+        'inline-flex items-center justify-center whitespace-nowrap rounded-sm px-3 py-1.5 text-sm font-medium ring-offset-background transition-all focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50',
+        isActive
+          ? 'bg-background text-foreground shadow-sm'
+          : 'hover:bg-background/50 hover:text-foreground',
+        className
+      )}
+      onClick={() => context.onValueChange(value)}
+      {...props}
+    >
+      {children}
+    </button>
+  );
+};
+
+export const TabsContent = ({ value, children, className, ...props }) => {
+  const context = useContext(TabsContext);
+  
+  if (!context) {
+    throw new Error('TabsContent must be used within Tabs');
+  }
+
+  if (context.value !== value) {
+    return null;
+  }
+
+  return (
+    <div
+      className={cn(
+        'mt-2 ring-offset-background focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2',
         className
       )}
       {...props}
-    />
-  )
-}
-
-function TabsContent({
-  className,
-  ...props
-}: React.ComponentProps<typeof TabsPrimitive.Content>) {
-  return (
-    <TabsPrimitive.Content
-      data-slot="tabs-content"
-      className={cn("flex-1 outline-none", className)}
-      {...props}
-    />
-  )
-}
-
-export { Tabs, TabsList, TabsTrigger, TabsContent }
+    >
+      {children}
+    </div>
+  );
+};
