+++ dashboard/frontend/src/contexts/AuthContext.tsx
