+++ dashboard/frontend/src/contexts/AuthContextDeclaration.tsx
