+++ dashboard/frontend/src/hooks/useAuth.ts
