--- dashboard/frontend/src/lib/utils.jsx
@@ -1,6 +1,0 @@
-import { clsx } from 'clsx';
-import { twMerge } from 'tailwind-merge';
-
-export function cn(...inputs) {
-  return twMerge(clsx(inputs));
-}
