+++ dashboard/frontend/vite.config.js
@@ -16,6 +16,6 @@ export default defineConfig({
   test: {
     environment: 'jsdom',
     globals: true,
-    setupFiles: './src/setupTests.jsx',
+    setupFiles: './src/setupTests.tsx',
   },
 })
